refactoring 08-01-2021

+++ src/main/java/com/epam/web/service/book/BookServiceImpl.java
@@ -3,7 +3,9 @@ package com.epam.web.service.book;
 import com.epam.web.dao.book.BookDao;
 import com.epam.web.dao.helper.DaoHelperImpl;
 import com.epam.web.dao.helper.DaoHelperFactory;
+import com.epam.web.dao.order.OrderDao;
 import com.epam.web.entity.Book;
+import com.epam.web.entity.Order;
 import com.epam.web.exception.ConnectionPoolException;
 import com.epam.web.exception.DaoException;
 import com.epam.web.exception.ServiceException;
@@ -19,16 +21,6 @@ public class BookServiceImpl implements BookService {
 
         this.daoHelperFactory = new DaoHelperFactory();
     }
-
-//    @Override
-//    public List<Book> findAll() throws ServiceException {
-//        try (DaoHelper daoHelper = daoHelperFactory.create()) {
-//            BookDao bookDao = daoHelper.createBookDao();
-//            return bookDao.findAllBooks();
-//        } catch (ConnectionPoolException | SQLException | DaoException e) {
-//            throw new ServiceException(e.getMessage(), e);
-//        }
-//    }
 
     @Override
     public List<Book> getBooksPart(int startPosition, int endPosition) throws ServiceException {
@@ -52,13 +44,36 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public void deleteBookById(Integer bookId) throws ServiceException {
+    public boolean deleteBookById(Integer bookId) throws ServiceException {
         try (DaoHelperImpl daoHelper = daoHelperFactory.create()) {
+            boolean result = false;
             BookDao bookDao = daoHelper.createBookDao();
-            bookDao.removeById(bookId);
+            OrderDao orderDao = daoHelper.createOrderDao();
+            List<Order> orders = orderDao.findAll();
+            if (!isBookInAcceptedOrders(bookId, orders)) {
+                //transaction
+                daoHelper.startTransaction();
+                bookDao.removeById(bookId);
+                orderDao.rejectPendingOrdersByBookId(bookId);
+                daoHelper.commitTransaction();
+
+                result = true;
+            }
+            return result;
         } catch (DaoException | ConnectionPoolException | SQLException e) {
             throw new ServiceException(e.getMessage(), e);
         }
+    }
+
+    private boolean isBookInAcceptedOrders(Integer bookId, List<Order> orders) {
+        boolean result = false;
+        for (Order order : orders) {
+            if (order.getBookId() == bookId && order.getStatus() == Order.Status.ACCEPTED) {
+                result = true;
+                break;
+            }
+        }
+        return result;
     }
 
     @Override
